refactor: js => ts
