+++ site/theme/template/Color/ColorPaletteTool.tsx
@@ -7,12 +7,15 @@ const primaryMinSaturation = 70; // 主色推荐最小饱和度
 const primaryMinBrightness = 70; // 主色推荐最小亮度
 
 export default class ColorPaletteTool extends Component {
-  state = {
+  state: {
+    primaryColor: string;
+    primaryColorInstance: any;
+  } = {
     primaryColor: '#1890ff',
     primaryColorInstance: null,
   };
 
-  handleChangeColor = (e, color) => {
+  handleChangeColor = (e: any, color: { hex: string }) => {
     const value = e.target ? e.target.value : e;
     this.setState({
       primaryColor: value,
